+++ R/mosaic3d.R
@@ -113,7 +113,7 @@
 #' @examples
 #'
 #' # 2 x 2 x 2
-#' if (!rgl.useNULL() && interactive()){
+#' if (!rgl::rgl.useNULL() && interactive()){
 #' mosaic3d(Bartlett, box=TRUE)
 #' # compare with expected frequencies under model of mutual independence
 #' mosaic3d(Bartlett, type="expected", box=TRUE)